+++ Moco/Examples/C++/exampleSlidingMass/exampleSlidingMass.cpp
@@ -100,7 +100,7 @@ int main() {
     // Configure the solver.
     // =====================
     MocoCasADiSolver& solver = moco.initCasADiSolver();
-    solver.set_num_mesh_points(50);
+    solver.set_num_mesh_intervals(50);
 
     // Now that we've finished setting up the tool, print it to a file.
     moco.print("sliding_mass.omoco");