+++ Moco/Moco/MocoGoal/MocoPeriodicityGoal.cpp
@@ -43,6 +43,7 @@ MocoPeriodicityGoalPair::MocoPeriodicityGoalPair(
 void MocoPeriodicityGoalPair::constructProperties() {
     constructProperty_initial_variable("");
     constructProperty_final_variable("");
+    constructProperty_negate(false);
 }
 
 //=============================================================================
@@ -63,15 +64,16 @@ void MocoPeriodicityGoal::initializeOnModelImpl(const Model& model) const {
 
     for (int i = 0; i < nStatePairs; ++i) {
         const auto path1 = get_state_pairs(i).get_initial_variable();
-        OPENSIM_THROW_IF(allSysYIndices.count(path1) == 0,
-                Exception, format("Could not find state '%s'.", path1));
+        OPENSIM_THROW_IF(allSysYIndices.count(path1) == 0, Exception,
+                format("Could not find state '%s'.", path1));
         const auto path2 = get_state_pairs(i).get_final_variable();
-        OPENSIM_THROW_IF(allSysYIndices.count(path2) == 0,
-                Exception, format("Could not find state '%s'.", path2));
+        OPENSIM_THROW_IF(allSysYIndices.count(path2) == 0, Exception,
+                format("Could not find state '%s'.", path2));
         int stateIndex1 = allSysYIndices[path1];
         int stateIndex2 = allSysYIndices[path2];
         m_state_names.emplace_back(path1, path2);
-        m_indices_states.emplace_back(stateIndex1, stateIndex2);
+        m_indices_states.emplace_back(stateIndex1, stateIndex2,
+                get_state_pairs(i).get_negate() ? -1 : 1);
     }
 
     auto systemControlIndexMap = createSystemControlIndexMap(model);
@@ -79,19 +81,20 @@ void MocoPeriodicityGoal::initializeOnModelImpl(const Model& model) const {
 
     for (int i = 0; i < nControlPairs; ++i) {
         const auto path1 = get_control_pairs(i).get_initial_variable();
-        OPENSIM_THROW_IF(systemControlIndexMap.count(path1) == 0,
-                Exception, format("Could not find control '%s'.", path1));
+        OPENSIM_THROW_IF(systemControlIndexMap.count(path1) == 0, Exception,
+                format("Could not find control '%s'.", path1));
         const auto path2 = get_control_pairs(i).get_final_variable();
-        OPENSIM_THROW_IF(systemControlIndexMap.count(path2) == 0,
-                Exception, format("Could not find control '%s'.", path2));
+        OPENSIM_THROW_IF(systemControlIndexMap.count(path2) == 0, Exception,
+                format("Could not find control '%s'.", path2));
         int controlIndex1 = systemControlIndexMap[path1];
         int controlIndex2 = systemControlIndexMap[path2];
+        m_indices_states.emplace_back(controlIndex1, controlIndex2,
+                get_state_pairs(i).get_negate() ? -1 : 1);
         m_control_names.emplace_back(path1, path2);
-        m_indices_controls.emplace_back(controlIndex1, controlIndex2);
     }
 
-    setNumIntegralsAndOutputs(0,
-            (int)m_indices_states.size() + (int)m_indices_controls.size());
+    setNumIntegralsAndOutputs(
+            0, (int)m_indices_states.size() + (int)m_indices_controls.size());
 }
 
 void MocoPeriodicityGoal::calcGoalImpl(
@@ -101,11 +104,10 @@ void MocoPeriodicityGoal::calcGoalImpl(
     const auto& finalStates = input.final_state.getY();
     int i = 0;
     for (const auto& index_state : m_indices_states) {
-        goal[i] = initialStates[index_state.first] -
-                finalStates[index_state.second];
-        if (getModeIsCost()) {
-            goal[i] = SimTK::square(goal[i]);
-        }
+        goal[i] = (initialStates[std::get<0>(index_state)] *
+                          std::get<2>(index_state)) -
+                  finalStates[std::get<1>(index_state)];
+        if (getModeIsCost()) { goal[i] = SimTK::square(goal[i]); }
         ++i;
     }
 
@@ -113,11 +115,10 @@ void MocoPeriodicityGoal::calcGoalImpl(
     const auto& finalControls = getModel().getControls(input.final_state);
     int j = 0;
     for (const auto& index_control : m_indices_controls) {
-        goal[i+j] = initialControls[index_control.first] -
-                finalControls[index_control.second];
-        if (getModeIsCost()) {
-            goal[i+j] = SimTK::square(goal[i + j]);
-        }
+        goal[i + j] = (initialControls[std::get<0>(index_control)] *
+                              std::get<2>(index_control)) -
+                      finalControls[std::get<1>(index_control)];
+        if (getModeIsCost()) { goal[i + j] = SimTK::square(goal[i + j]); }
         ++j;
     }
 }