+++ Moco/Tests/testDeGrooteFregly2016Muscle.cpp
@@ -616,7 +616,7 @@ TEMPLATE_TEST_CASE("Hanging muscle minimum time", "", MocoCasADiSolver) {
         problem.addGoal<MocoFinalTimeGoal>();
 
         auto& solver = moco.initSolver<TestType>();
-        solver.set_num_mesh_points(40);
+        solver.set_num_mesh_intervals(40);
         solver.set_dynamics_mode("implicit");
         solver.set_optim_convergence_tolerance(1e-4);
         solver.set_optim_constraint_tolerance(1e-3);
@@ -705,7 +705,7 @@ TEMPLATE_TEST_CASE("Hanging muscle minimum time", "", MocoCasADiSolver) {
         tracking->setAllowUnusedReferences(true);
 
         auto& solver = moco.initSolver<TestType>();
-        solver.set_num_mesh_points(40);
+        solver.set_num_mesh_intervals(40);
         solver.set_dynamics_mode("implicit");
         solver.set_transcription_scheme("trapezoidal");