+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+- 2019-08-29: Users now interface with num_mesh_intervals instead of
+              num_mesh_points when setting times to sample.
+
+- 2019-08-20: Periodicity of the states and/or controls can now be enforced with
+              a negated MocoPeriodicityGoalPair added via addNegatedStatePair or
+              addNegatedControlPair respectively.
+
 - 2019-07-18: Periodicity of the states and/or controls can be enforced with
               MocoPeriodicityGoals that impose equality of the initial and final
               state/control values.